Add player image files

+++ src/Command/PlayerImportCommand.php
@@ -6,10 +6,12 @@ use App\Entity\Player;
 use App\Entity\Position;
 use App\Entity\Team;
 use App\Provider\KickerProvider;
+use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
@@ -36,6 +38,7 @@ class PlayerImportCommand extends Command
     {
         $this
             ->setName('soc:player:import')
+            ->addOption('erase-data', null, InputOption::VALUE_OPTIONAL, 'Should the database data be erased?', false)
             ->setDescription('Imports Player from the Kicker.de Website in Storage')
             ->setHelp(
                 <<<EOT
@@ -44,39 +47,49 @@ class PlayerImportCommand extends Command
     Options:
     - clear
 EOT
-            );
+            )
+            ;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
 
-        $connection = $this->doctrine->getConnection();
+        $eraseData = $input->getOption('erase-data');
 
-        $connection->executeQuery('SET foreign_key_checks = 0');
-        $connection->executeQuery('TRUNCATE soc_player');
-        $connection->executeQuery('TRUNCATE soc_team');
-        $connection->executeQuery('SET foreign_key_checks = 1');
+        if (true === $eraseData) {
+            $connection = $this->doctrine->getConnection();
 
-        $this->updateTeams();
-        $io->success('Teams have been updated');
+            $connection->executeQuery('SET foreign_key_checks = 0');
+            $connection->executeQuery('TRUNCATE soc_player');
+            $connection->executeQuery('TRUNCATE soc_team');
+            $connection->executeQuery('SET foreign_key_checks = 1');
+        }
 
-        $this->updatePlayers();
-        $io->success('Players have been updated');
+        $this->updateTeams($io);
+        $this->updatePlayers($io);
 
         $io->success('Done.');
 
         return Command::SUCCESS;
     }
 
-    private function updatePlayers(): void
+    private function updatePlayers(SymfonyStyle $io): void
     {
         $om = $this->doctrine->getManager();
 
         $teamRepo = $om->getRepository(Team::class);
         $teams = $teamRepo->findAll();
+
         $positionRepo = $om->getRepository(Position::class);
         $positions = $positionRepo->findAll();
+
+        $playerRepo = $om->getRepository(Player::class);
+
+        $playersCollection = new ArrayCollection($playerRepo->findAll());
+
+        $playersUpdated = 0;
+        $playersCreated = 0;
 
         foreach ($positions as $position) {
             $players = $this->kickerProvider->getPlayersByPosition($position);
@@ -88,6 +101,16 @@ EOT
                     if ($team->getName() === $player['team']) {
                         $t = $team;
                     }
+                }
+
+                $player['externalId'] = (int) $player['externalId'];
+
+                if ($playersCollection->exists(function ($key, $element) use ($player) {
+                    return $element->getExternalId() === $player['externalId'];
+                })) {
+                    ++$playersUpdated;
+
+                    continue;
                 }
 
                 $entity = new Player();
@@ -105,12 +128,17 @@ EOT
                     ->setAverageScore($player['averageScore']);
 
                 $om->persist($entity);
+
+                ++$playersCreated;
             }
             $om->flush();
         }
+
+        $io->success("#$playersCreated Players have been created");
+        $io->success("#$playersUpdated Players have been updated");
     }
 
-    private function updateTeams(): void
+    private function updateTeams(SymfonyStyle $io): void
     {
         $om = $this->doctrine->getManager();
 
@@ -134,5 +162,7 @@ EOT
             $om->persist($team);
         }
         $om->flush();
+
+        $io->success('Teams have been updated');
     }
 }
